Introduce new setup command to consolidate artifacts to a namespace
Introduce a new `restructure` command that mangles the repo. This moves
the protobufs to directories inside the min_tfs_client namespace and
then does a string replace on all the protobuf imports to reflect the
new top level namespace.

This allows the package to provide the same, lightweight python runtime
as it was originally intended, while still being used in an image with
tensorflow installed.