Add import fallbacks to play nice if Tensorflow already loaded
Protobufs register themselves to a global namespace. Because of this, we
need to be defensive in loading the protobufs. Change adds try/excepts
for the importing of the protobufs, allowing the library to fall back on
the main tensorflow protobufs. Downside is that importing tensorflow
after this library will still throw the protobuf registration error